# web/swagger.yml
openapi: 3.0.0
info:
  title: Content Addressable Storage (CAS) Server API
  description: A file storage server that stores and retrieves files based on their SHA256 hash
  version: 1.0.0
servers:
  - url: /
    description: Current server
...
paths:
...
  /file/upload:
    post:
      tags:
        - casd
        - casd-balancer
      summary: Upload a file to CAS storage
      description: Uploads a file and stores it using its SHA256 hash. Returns conflict if file already exists.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
              required:
                - file
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  hash:
                    type: string
                    description: SHA256 hash of the uploaded file
                    example: "a1fff0ffefb9eace7230c24e50731f0a91c62f9cefdfe77121c2f607125dffae"
        '400':
          description: Bad request - file parameter is missing
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "file parameter is required"
        '409':
          description: Conflict - file already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "file already exists"
                  hash:
                    type: string
                    description: SHA256 hash of the existing file
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...